+++ app/src/main/java/com/trekkstay/hotel/feature/reservation/domain/entities/reservation.kt
@@ -8,11 +8,11 @@ data class Reservation(
     val roomId:String,
     val userId: String,
     val qrCodeUrl: String,
-    val quantity:Int,
+    val quantity:Double,
     val totalPrice: Double,
     val checkIn:String,
     val checkOut:String,
     val status: String,
     val guestInfo: GuestInfo,
     val room: ReservationRoom,
-    )
+)
